minor: updates database models (v3.2.0)
- `User` - replaces bio field with status
- `Note` - converts articles collection to notes and turns tag objects to strings
- `Tag` - removes collection
- `Article` - removes collection
- `Experience` - renames projects to project in type field
- updates corresponding Frontend references

+++ backend/src/database.ts
@@ -1,26 +1,35 @@
 import { DB_URI } from "./config/env";
 import mongoose from "mongoose";
+import { BaseError } from "./types/error";
 
-const connection = mongoose.connection;
+let connection: mongoose.Connection | undefined = undefined;
 
 /**
  * Connect to database
  */
 async function connect(): Promise<typeof mongoose> {
-	return await mongoose.connect(DB_URI);
+	const database = await mongoose.connect(DB_URI);
+  connection = database.connection;
+
+  return database;
 }
 
 /**
  * Disconnect from database
  */
 async function disconnect() {
-	return await mongoose.disconnect();
+	await mongoose.disconnect();
+  connection = undefined;
 }
 
 /**
  * Drop database
  */
 async function drop () {
+  if(!connection) {
+    throw { status: 500, message: "no connection to database" } as BaseError;
+  }
+
 	return connection.dropDatabase();
 }
